Solver::*Alternate* interface. Finally ran the full sudoku17.txt suite
Create a way of describing the problem with separate collections of filters.
During solving, the separate collections are prepared and their selectivity is used
to choose the form. This allows Grid problems to try both the default and the
transpose form of the problem, and this appears to cut down the worst behavior
of sudoku processing. Most solutions are 30ms with worst case comming in at 100-200ms, whereas >1s solutions where common prior.